Wisely 0.0.3 Antenna trace correction
Had to correct the antenna trace.  The antenna pin 1 is not labeled as a ground pin.  This should be corrected in the future because the application notes require that both the input pin and the ground pin connect to ground.  Sincedesign does not label the input pin as ground, KiCAD will disconnect it when refreshing the PCB layout. This commit reconnects the trace after the past commit had a break from the input pin to ground.